backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.10

# Cache
redis==5.0.1

# Ollama Integration
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

